# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Web API для управления шахматными партиями и сущностями (игроки, фигуры, ходы).",
        "title": "Chess Game API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/game": {
            "post": {
                "description": "Создает новую шахматную партию между двумя игроками",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "game"
                ],
                "summary": "Новая игра",
                "parameters": [
                    {
                        "description": "Параметры новой игры",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/packages_internal_handlers.createGame"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/packages_internal_handlers.response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/packages_internal_handlers.Err"
                        }
                    }
                }
            }
        },
        "/game/{id}": {
            "get": {
                "description": "Возвращает текущее состояние доски в JSON или HTML (зависит от Accept-заголовка)",
                "produces": [
                    "application/json",
                    "text/html"
                ],
                "tags": [
                    "game"
                ],
                "summary": "Отображение доски",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID игры",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/game/{id}/automove": {
            "post": {
                "description": "Запускает указанное количество автоматических ходов от лица игрока",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "game"
                ],
                "summary": "Автоход",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID игры",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Параметры автоматических ходов",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/packages_internal_handlers.autoMove"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/packages_internal_handlers.response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/packages_internal_handlers.Err"
                        }
                    }
                }
            }
        },
        "/game/{id}/move": {
            "post": {
                "description": "Выполняет ход указанного игрока в активной партии",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "game"
                ],
                "summary": "Сделать ход",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID игры",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Параметры хода",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/packages_internal_handlers.move"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/packages_internal_handlers.response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/packages_internal_handlers.Err"
                        }
                    }
                }
            }
        },
        "/game/{id}/stop": {
            "post": {
                "description": "Останавливает партию без объявления победителя",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "game"
                ],
                "summary": "Остановить игру",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID игры",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/packages_internal_handlers.response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/packages_internal_handlers.Err"
                        }
                    }
                }
            }
        },
        "/game/{id}/surrender": {
            "post": {
                "description": "Игрок сдается, партия завершается победой соперника",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "game"
                ],
                "summary": "Сдаться",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID игры",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Кто сдается",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/packages_internal_handlers.surrender"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/packages_internal_handlers.response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/packages_internal_handlers.Err"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "packages_internal_handlers.Err": {
            "type": "object",
            "properties": {
                "err": {
                    "type": "string"
                }
            }
        },
        "packages_internal_handlers.autoMove": {
            "type": "object",
            "properties": {
                "moves_amount": {
                    "type": "integer"
                },
                "player": {
                    "type": "string"
                }
            }
        },
        "packages_internal_handlers.createGame": {
            "type": "object",
            "properties": {
                "board_size": {
                    "type": "integer"
                },
                "player_one_name": {
                    "type": "string"
                },
                "player_two_name": {
                    "type": "string"
                }
            }
        },
...
        "packages_internal_handlers.move": {
            "type": "object",
            "properties": {
                "player": {
                    "type": "string"
                },
                "position_from": {
                    "type": "string"
                },
                "position_to": {
                    "type": "string"
                }
            }
        },
...
        "packages_internal_handlers.response": {
            "type": "object",
            "properties": {
                "game_id": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "packages_internal_handlers.surrender": {
            "type": "object",
            "properties": {
                "player": {
                    "type": "string"
                }
            }
        },
...
    }
}